+++ isense.h
@@ -3,6 +3,6 @@
 
 void ADC_init();
 unsigned int ADC_Sample_Convert();
-unsigned int ADC_mA();
+float ADC_mA();
 
 #endif
